+++ examples/src/simulflow_examples/gemini.clj
@@ -6,24 +6,18 @@
    [simulflow.async :refer [vthread-loop]]
    [simulflow.processors.google :as google]
    [simulflow.secrets :refer [secret]]
+   [simulflow.vad.silero :as silero]
    [taoensso.telemere :as t]))
 
-(def local-flow-gemini
-  (local/make-local-flow {:extra-procs {:llm {:proc google/google-llm-process
-                                              :args {:llm/model :gemini-2.0-flash
-                                                     :google/api-key (secret [:google :api-key])}}}
-                          :llm-context {:messages
-                                        [{:role "system"
-                                          :content "You are a voice agent operating via phone. Be
-                       concise in your answers. The input you receive comes from a
-                       speech-to-text (transcription) system that isn't always
-                       efficient and may send unclear text. Ask for
-                       clarification when you're unsure what the person said."}]
-                                        :tools []}}))
-
-(defonce flow-started? (atom false))
-
 (comment
+  (def local-flow-gemini
+    (local/make-local-flow {:extra-procs {:llm {:proc google/google-llm-process
+                                                :args {:llm/model :gemini-2.0-flash
+                                                       :google/api-key (secret [:google :api-key])}}}
+                            :vad-analyser (silero/create-silero-vad)}))
+
+  (defonce flow-started? (atom false))
+
   ;; Start local ai flow - starts paused
   (let [{:keys [report-chan error-chan]} (flow/start local-flow-gemini)]
     (reset! flow-started? true)