+++ src/simulflow/processors/google.clj
@@ -1,18 +1,12 @@
 (ns simulflow.processors.google
   (:require
-   [clojure.core.async :as a]
    [clojure.core.async.flow :as flow]
    [clojure.string :as str]
    [hato.client :as http]
-   [malli.core :as m]
-   [malli.error :as me]
-   [simulflow.async :refer [vthread-loop]]
-   [simulflow.frame :as frame]
    [simulflow.schema :as schema]
    [simulflow.secrets :refer [secret]]
    [simulflow.utils.core :as u]
-   [simulflow.utils.openai :as uai]
-   [taoensso.telemere :as t]))
+   [simulflow.utils.openai :as uai]))
 
 (def google-generative-api-url "https://generativelanguage.googleapis.com/v1beta/openai")
 (def google-completions-url (str google-generative-api-url "/chat/completions"))
@@ -33,23 +27,26 @@
   (schema/flex-enum
     {:description "Google llm model identifier"
      :error/message "Must be a valid Google LLM model. Try gemini-2.0-flash"}
-    ["gemini-embedding-exp" "gemini-2.0-flash-thinking-exp" "learnlm-1.5-pro-experimental" "gemini-1.5-pro" "gemini-2.0-flash-lite-001" "gemini-1.5-flash"
-     "gemini-2.0-flash-lite-preview" "chat-bison-001" "gemini-exp-1206" "gemini-1.5-pro-002" "text-bison-001" "gemini-2.0-flash-lite-preview-02-05"
-     "gemini-1.5-flash-8b-latest" "embedding-001" "gemini-2.0-pro-exp" "gemini-1.5-flash-8b-001" "gemini-1.0-pro-vision-latest" "gemini-2.5-pro-preview-03-25"
-     "learnlm-2.0-flash-experimental" "gemini-1.5-flash-002" "gemma-3-27b-it" "text-embedding-004" "gemini-1.5-flash-001-tuning" "gemini-2.0-flash-lite"
-     "gemini-2.5-flash-preview-04-17-thinking" "gemini-1.5-pro-001" "gemini-pro-vision" "gemini-1.5-flash-8b-exp-0924" "gemini-2.0-flash-live-001" "aqa"
-     "gemini-1.5-flash-8b" "gemini-2.0-flash-thinking-exp-1219" "gemini-1.5-flash-latest" "gemini-1.5-pro-latest" "gemma-3-4b-it" "embedding-gecko-001"
-     "gemini-2.0-flash-thinking-exp-01-21" "gemini-2.0-pro-exp-02-05" "veo-2.0-generate-001" "gemini-embedding-exp-03-07" "gemini-2.5-pro-exp-03-25" "gemini-1.5-flash-8b-exp-0827"
-     "gemma-3-12b-it" "gemini-2.0-flash" "gemini-2.5-flash-preview-04-17" "gemini-2.0-flash-001" "gemini-1.5-flash-001" "gemma-3-1b-it" "gemini-2.5-pro-preview-05-06"
-     "imagen-3.0-generate-002" "gemini-2.0-flash-exp"]))
+    [:gemini-embedding-exp :gemini-2.0-flash-thinking-exp :learnlm-1.5-pro-experimental :gemini-1.5-pro :gemini-2.0-flash-lite-001 :gemini-1.5-flash
+     :gemini-2.0-flash-lite-preview :chat-bison-001 :gemini-exp-1206 :gemini-1.5-pro-002 :text-bison-001 :gemini-2.0-flash-lite-preview-02-05
+     :gemini-1.5-flash-8b-latest :embedding-001 :gemini-2.0-pro-exp :gemini-1.5-flash-8b-001 :gemini-1.0-pro-vision-latest :gemini-2.5-pro-preview-03-25
+     :learnlm-2.0-flash-experimental :gemini-1.5-flash-002 :gemma-3-27b-it :text-embedding-004 :gemini-1.5-flash-001-tuning :gemini-2.0-flash-lite
+     :gemini-2.5-flash-preview-04-17-thinking :gemini-1.5-pro-001 :gemini-pro-vision :gemini-1.5-flash-8b-exp-0924 :gemini-2.0-flash-live-001
+     :gemini-1.5-flash-8b :gemini-2.0-flash-thinking-exp-1219 :gemini-1.5-flash-latest :gemini-1.5-pro-latest :gemma-3-4b-it :embedding-gecko-001
+     :gemini-2.0-flash-thinking-exp-01-21 :gemini-2.0-pro-exp-02-05 :veo-2.0-generate-001 :gemini-embedding-exp-03-07 :gemini-2.5-pro-exp-03-25 :gemini-1.5-flash-8b-exp-0827
+     :gemma-3-12b-it :gemini-2.0-flash :gemini-2.5-flash-preview-04-17 :gemini-2.0-flash-001 :gemini-1.5-flash-001 :gemma-3-1b-it :gemini-2.5-pro-preview-05-06
+     :imagen-3.0-generate-002 :gemini-2.0-flash-exp]))
 
 (def GoogleLLMConfigSchema
   [:map
    {:description "Google LLM configuration"}
-   [:llm/model {:default :gemini-2.0-flash} model-schema]
-   [:google/api-key [:string
-                     {:description "Google API key"
-                      :error/message "Invalid Google Api Key provided"}]]])
+   [:llm/model {:default :gemini-2.0-flash
+                :description "Google model used for llm inference"} model-schema]
+   [:google/api-key {:description "Google API key"
+                     :error/message "Invalid Google Api Key provided"} :string]
+   [:api/completions-url {:default google-completions-url
+                          :optional true
+                          :description "Different completions url for gemini api"} :string]])
 
 (comment
 
@@ -64,48 +61,29 @@
                                                  {:role "user" :content "Do you hear me?"}]
                                       :completions-url google-completions-url})))
 
-(defn google-llm-process-fn
-  ([]
-   {:ins {:in "Channel for incoming context aggregations"}
-    :outs {:out "Channel where streaming responses will go"}
-    :params (schema/->describe-parameters GoogleLLMConfigSchema)
-    :workload :io})
-  ([params]
-   (when-let [error (m/explain GoogleLLMConfigSchema params)]
-     (throw (ex-info "google-llm-processor: Invalid configuration"
-                     {:humanized (me/humanize error)
-                      :error error
-                      :type ::invalid-configuration})))
-   (let [llm-write (a/chan 1024)
-         llm-read (a/chan 1024)
-         {:llm/keys [model] :google/keys [api-key]} params]
-     (vthread-loop []
-                   (when-let [frame (a/<!! llm-write)]
-                     (t/log! {:level :info :id :google} ["Processing request" (:frame/data frame)])
-                     (assert (or (frame/llm-context? frame)
-                                 (frame/control-interrupt-start? frame)) "Invalid frame sent to LLM. Only llm-context or interrupt-start")
-                     (let [context (:frame/data frame)
-                           stream-ch (uai/stream-chat-completion {:model model
-                                                                  :api-key api-key
-                                                                  :messages (:messages context)
-                                                                  :tools (mapv u/->tool-fn (:tools context))
-                                                                  :completions-url google-completions-url})]
-                       (uai/handle-completion-request! stream-ch llm-read))
+(def describe
+  {:ins {:in "Channel for incoming context aggregations"}
+   :outs {:out "Channel where streaming responses will go"}
+   :params (schema/->describe-parameters GoogleLLMConfigSchema)
+   :workload :io})
 
-                     (recur)))
-     {::flow/in-ports {:llm-read llm-read}
-      ::flow/out-ports {:llm-write llm-write}}))
-  ([{::flow/keys [in-ports out-ports] :as state} transition]
-   (when (= transition ::flow/stop)
-     (doseq [port (concat (vals in-ports) (vals out-ports))]
-       (a/close! port)))
-   state)
-  ([state in msg]
-   (if (= in :llm-read)
-     [state {:out [msg]}]
-     (cond
-       (frame/llm-context? msg)
-       [state {:llm-write [msg]
-               :out [(frame/llm-full-response-start true)]}]))))
+(defn init!
+  [params]
+  (uai/init-llm-processor! GoogleLLMConfigSchema params :gemini))
+
+(defn transition
+  [state transition]
+  (uai/transition-llm-processor state transition))
+
+(defn transform
+  [state in msg]
+  (uai/transform-llm-processor state in msg :google/api-key :api/completions-url))
+
+(defn google-llm-process-fn
+  ([] describe)
+  ([params] (init! params))
+  ([state trs]
+   (transition state trs))
+  ([state in msg] (transform state in msg)))
 
 (def google-llm-process (flow/process google-llm-process-fn))